+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 As this project is still in active development, it does not yet strictly adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `skip_evaluation` for classifying unlabeled data (optional/empty `flag`); skips label creation and performance evaluation
 
 ## [0.3.1] - 2026-07-03
 ### Fixed